+++ src/hkimjp/wil2/todays.clj
@@ -1,8 +1,10 @@
 (ns hkimjp.wil2.todays
   (:require
    [hkimjp.datascript :as ds]
+   [hkimjp.wil2.util :refer [user]]
    [hkimjp.wil2.view :refer [page]]
    [java-time.api :as jt]
+   [ring.util.response :as resp]
    [taoensso.telemere :as t]))
 
 (def uploaded? '[:find ?e
@@ -18,12 +20,16 @@
 
 (comment
   (ds/qq uploaded? "hkimura")
+  (some? (first (ds/qq uploaded? "hkimura")))
+  (ds/qq uploaded? "chatgpt")
+  (some? (first (ds/qq uploaded? "chatgpt")))
   (ds/qq todays-uploads "2025-09-04")
   (def data {:login "hkimura"
              :md "# hello, World"
              :date "2025-09-04"
              :updated (jt/local-date-time)})
   (ds/put! data)
+  (:md (ds/pl 1))
   :rcf)
 
 (defn upload [request]
@@ -36,7 +42,11 @@
   (page [:div "todays"]))
 
 (defn switch [request]
-  (page
-   [:div.text-2xl "switch"
-    [:div [:a {:href "/wil2/upload"} "upload"]]
-    [:div [:a {:href "/wil2/todays"} "todays"]]]))
+  (if (some? (first (ds/qq uploaded? (user request))))
+    (resp/redirect "/wil2/todays")
+    (resp/redirect "/wil2/upload")))
+
+; (page
+  ;  [:div.text-2xl "switch"
+  ;   [:div [:a {:href "/wil2/upload"} "upload"]]
+  ;   [:div [:a {:href "/wil2/todays"} "todays"]]]))